kisaan-llm-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
llama-cpp-python>=0.3.0
huggingface_hub>=0.23.0
python-multipart>=0.0.9
pydantic>=2.0.0
